Added Notifications, Middleware

+++ routes/web.php
@@ -32,57 +32,85 @@ Auth::routes();
 //[REMOVE THIS CODE IT JUST FOR TESTING EMAILS]	
 Route::get('/sendmail', function(){
 
-		$user = User::find(1);
+	$user = User::find(1);
 
-		echo 'Preparing to send...';
+	echo 'Preparing to send...';
 
-        //Mail user to activate account 
-        Mail::to('brandontabona@gmail.com')->send(new AccountActivated($user));
+    //Mail user to activate account 
+    Mail::to('brandontabona@gmail.com')->send(new AccountActivated($user));
 
-        echo 'Email sent successfully';
+    echo 'Email sent successfully';
 
 });
 
-Route::group(['prefix' => 'messages'], function () {
-    Route::get('/', ['as' => 'messages', 'uses' => 'MessagesController@index']);
-    Route::get('create', ['as' => 'messages.create', 'uses' => 'MessagesController@create']);
-    Route::post('/', ['as' => 'messages.store', 'uses' => 'MessagesController@store']);
-    Route::get('{id}', ['as' => 'messages.show', 'uses' => 'MessagesController@show']);
-    Route::put('{id}', ['as' => 'messages.update', 'uses' => 'MessagesController@update']);
+Route::get('/error', function(){
+
+	return view('error.connection_fail');
+
+}); 
+
+Route::get('/support', function(){
+
+	return view('/support');
+
+});
+
+Route::get('/sendsms', function(){
+	
+	echo 'Preparing to send...';
+
+	Nexmo::message()->send([
+	    'to'   => '26775993221',
+	    'from' => '26776710242',
+	    'text' => 'Hi Julian, Our Saleschief team has sent you an email to enroll into our sales class. See you there :)'
+	]);
+
+    echo 'SMS sent successfully';
+
+});
+
+Route::middleware(['auth'])->group(function () {
+	Route::group(['prefix' => 'messages'], function () {
+	    Route::get('/', ['as' => 'messages', 'uses' => 'MessagesController@index']);
+	    Route::get('create', ['as' => 'messages.create', 'uses' => 'MessagesController@create']);
+	    Route::post('/', ['as' => 'messages.store', 'uses' => 'MessagesController@store']);
+	    Route::get('{id}', ['as' => 'messages.show', 'uses' => 'MessagesController@show']);
+	    Route::put('{id}', ['as' => 'messages.update', 'uses' => 'MessagesController@update']);
+	});
 });
 
 
 Route::get('/courses', 'CourseController@index')->name('course-list');
-Route::post('/courses', 'CourseController@store')->name('course-save');
-Route::get('/courses/enroll', 'CourseController@enroll')->name('course-enroll');
-Route::get('/courses/create', 'CourseController@create')->name('course-create');
-Route::get('/courses/{course_id}', 'CourseController@show');
-Route::put('/courses/{course_id}', 'CourseController@update');
-Route::delete('/courses/{course_id}', 'CourseController@delete');
-Route::get('/courses/{course_id}/edit', 'CourseController@edit');
-Route::get('/courses/{course_id}/clients/create', 'UserController@enroll')->name('client-enroll');
+Route::post('/courses', 'CourseController@store')->name('course-save')->middleware('onlyadmin');
+Route::get('/courses/enroll', 'CourseController@enroll')->name('course-enroll')->middleware('onlyadmin');
+Route::get('/courses/create', 'CourseController@create')->name('course-create')->middleware('onlyadmin');
+Route::get('/courses/{course_id}', 'CourseController@show')->middleware('auth');
+Route::put('/courses/{course_id}', 'CourseController@update')->middleware('onlyadmin');
+Route::delete('/courses/{course_id}', 'CourseController@delete')->middleware('onlyadmin');
+Route::get('/courses/{course_id}/edit', 'CourseController@edit')->middleware('onlyadmin');
+Route::get('/courses/{course_id}/clients/create', 'UserController@enroll')->name('client-enroll')->middleware('onlyadmin');
 
 
 /* 
 	MODULES - ADD, EDIT, UPDATE, DELETE
 */
 
-Route::post('/courses/{id}/module', 'ModuleController@store');
-Route::put('/courses/{course_id}/module/{module_id}', 'ModuleController@update');
-Route::delete('/courses/{course_id}/module/{module_id}', 'ModuleController@delete');
-Route::get('/courses/{course_id}/module/{module_id}/edit', 'ModuleController@edit');
+Route::post('/courses/{id}/module', 'ModuleController@store')->middleware('onlyadmin');
+Route::put('/courses/{course_id}/module/{module_id}', 'ModuleController@update')->middleware('onlyadmin');
+Route::delete('/courses/{course_id}/module/{module_id}', 'ModuleController@delete')->middleware('onlyadmin');
+Route::get('/courses/{course_id}/module/{module_id}/edit', 'ModuleController@edit')->middleware('onlyadmin');
 
 /* 
 	LESSONS - ADD, EDIT, UPDATE, DELETE
 */
-Route::post('/courses/{course_id}/module/{module_id}/lesson/notes', 'LessonController@notesImageUpload');
-Route::post('/courses/{course_id}/module/{module_id}/lesson', 'LessonController@store');
-Route::get('/courses/{course_id}/module/{module_id}/lesson/create', 'LessonController@create');
-Route::put('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}', 'LessonController@update');
-Route::delete('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}', 'LessonController@delete');
-Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/edit', 'LessonController@edit');
-Route::post('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/video', 'LessonController@videoUpload')->name('lesson-video-save');
-Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/video', 'LessonController@video')->name('lesson-video');
+Route::post('/courses/{course_id}/module/{module_id}/lesson/notes', 'LessonController@notesImageUpload')->middleware('onlyadmin');
+Route::post('/courses/{course_id}/module/{module_id}/lesson', 'LessonController@store')->middleware('onlyadmin');
+Route::get('/courses/{course_id}/module/{module_id}/lesson/create', 'LessonController@create')->middleware('onlyadmin');
+Route::put('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}', 'LessonController@update')->middleware('onlyadmin');
+Route::delete('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}', 'LessonController@delete')->middleware('onlyadmin');
+Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/edit', 'LessonController@edit')->middleware('onlyadmin');
+Route::post('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/video', 'LessonController@videoUpload')->name('lesson-video-save')->middleware('onlyadmin');
+Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/video', 'LessonController@video')->name('lesson-video')->middleware('onlyadmin');
 
 /* 
 	LESSON NOTES - ADD, EDIT, UPDATE, DELETE
@@ -94,13 +122,13 @@ Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/video', '
 	TESTS - ADD, EDIT, UPDATE, DELETE
 */
 Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests', 'TestController@index')->name('test-list');
-Route::post('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests', 'TestController@store')->name('test-save');
-Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/create', 'TestController@create')->name('test-create');
-Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}', 'TestController@taketest');
-Route::put('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}', 'TestController@update');
-Route::delete('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}', 'TestController@delete');
-Route::post('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}/mark', 'TestController@markTest');
-Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}/edit', 'TestController@edit');
+Route::post('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests', 'TestController@store')->name('test-save')->middleware('onlyadmin');
+Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/create', 'TestController@create')->name('test-create')->middleware('onlyadmin');
+Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}', 'TestController@taketest')->middleware('auth');
+Route::put('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}', 'TestController@update')->middleware('onlyadmin');
+Route::delete('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}', 'TestController@delete')->middleware('onlyadmin');
+Route::post('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}/mark', 'TestController@markTest')->middleware('auth');
+Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{test_id}/edit', 'TestController@edit')->middleware('onlyadmin');
 
 //markTtest
 //mark
@@ -109,31 +137,31 @@ Route::get('/courses/{course_id}/module/{module_id}/lesson/{lesson_id}/tests/{te
 	CLIENTS - ADD, EDIT, UPDATE, DELETE
 */
 
-Route::get('/clients', 'UserController@index')->name('client-list');
-Route::post('/clients', 'UserController@store')->name('client-save');
-Route::get('/clients/create', 'UserController@create')->name('client-create');
-Route::get('/clients/{client_id}', 'UserController@show')->name('client-update');
-Route::put('/clients/{client_id}', 'UserController@update');
-Route::get('/clients/account/setup/{client_email}', 'UserController@setup')->name('client-setup');
+Route::get('/clients', 'UserController@index')->name('client-list')->middleware('onlyadmin');
+Route::post('/clients', 'UserController@store')->name('client-save')->middleware('onlyadmin');
+Route::get('/clients/create', 'UserController@create')->name('client-create')->middleware('onlyadmin');
+Route::get('/clients/{client_id}', 'UserController@show')->name('client-update')->middleware('onlyadmin');
+Route::put('/clients/{client_id}', 'UserController@update')->middleware('onlyadmin');
+Route::get('/clients/account/setup/{client_email}', 'UserController@setup')->name('client-setup')->middleware('guest');
 Route::get('/clients/activate/{client_email}/{client_token}', 'UserController@activate')->name('client-activate');
 
 /* 
 	COMPANIES - ADD, EDIT, UPDATE, DELETE
 */
 
-Route::get('/companies', 'CompanyController@index')->name('company-list');
-Route::post('/companies', 'CompanyController@store')->name('company-save');
-Route::get('/companies/create', 'CompanyController@create')->name('company-create');
-Route::get('/companies/{company_id}', 'CompanyController@show')->name('company-show');
-Route::put('/companies/{company_id}', 'CompanyController@update');
-Route::delete('/companies/{company_id}', 'CompanyController@delete');
-Route::get('/companies/{company_id}/edit', 'CompanyController@edit');
+Route::get('/companies', 'CompanyController@index')->name('company-list')->middleware('onlyadmin');
+Route::post('/companies', 'CompanyController@store')->name('company-save')->middleware('onlyadmin');
+Route::get('/companies/create', 'CompanyController@create')->name('company-create')->middleware('onlyadmin');
+Route::get('/companies/{company_id}', 'CompanyController@show')->name('company-show')->middleware('onlyadmin');
+Route::put('/companies/{company_id}', 'CompanyController@update')->middleware('onlyadmin');
+Route::delete('/companies/{company_id}', 'CompanyController@delete')->middleware('onlyadmin');
+Route::get('/companies/{company_id}/edit', 'CompanyController@edit')->middleware('onlyadmin');
 
 /* 
 	ADMINS - ADD, EDIT, UPDATE, DELETE
 */
 
-Route::get('/admins/{admin_id}', 'AdminController@show')->name('admin-show');
+Route::get('/admins/{admin_id}', 'AdminController@show')->name('admin-show')->middleware('onlyadmin');
 
 /* 
 	SEARCH - ADD, EDIT, UPDATE, DELETE
